Created a more inheritable hierarchy of classes
