+++ rules/downgrade/src/Rector/FunctionLike/AbstractDowngradeParamTypeDeclarationRector.php
@@ -12,23 +12,19 @@ use PhpParser\Node\Param;
 use PhpParser\Node\Stmt\ClassMethod;
 use PhpParser\Node\Stmt\Function_;
 use Rector\BetterPhpDocParser\PhpDocInfo\PhpDocInfo;
-use Rector\Core\Contract\Rector\ConfigurableRectorInterface;
-use Rector\Downgrade\Contract\Rector\DowngradeRectorInterface;
 use Rector\Downgrade\Contract\Rector\DowngradeTypeRectorInterface;
+use Rector\Downgrade\Rector\Property\AbstractDowngradeRector;
 use Rector\NodeTypeResolver\Node\AttributeKey;
-use Rector\TypeDeclaration\Rector\FunctionLike\AbstractTypeDeclarationRector;
 
-abstract class AbstractDowngradeParamTypeDeclarationRector extends AbstractTypeDeclarationRector implements ConfigurableRectorInterface, DowngradeRectorInterface, DowngradeTypeRectorInterface
+abstract class AbstractDowngradeParamTypeDeclarationRector extends AbstractDowngradeRector implements DowngradeTypeRectorInterface
 {
     /**
-     * @var string
+     * @return string[]
      */
-    public const ADD_DOC_BLOCK = '$addDocBlock';
-
-    /**
-     * @var bool
-     */
-    private $addDocBlock = true;
+    public function getNodeTypes(): array
+    {
+        return [Function_::class, ClassMethod::class];
+    }
 
     /**
      * @param ClassMethod|Function_ $node
@@ -48,11 +44,6 @@ abstract class AbstractDowngradeParamTypeDeclarationRector extends AbstractTypeD
         }
 
         return null;
-    }
-
-    public function configure(array $configuration): void
-    {
-        $this->addDocBlock = $configuration[self::ADD_DOC_BLOCK] ?? true;
     }
 
     protected function getRectorDefinitionDescription(): string
