+++ rules/downgrade/src/Rector/Property/AbstractDowngradeRector.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Downgrade\Rector\Property;
+
+use Rector\Downgrade\Contract\Rector\DowngradeRectorInterface;
+
+abstract class AbstractDowngradeRector extends AbstractMaybeAddDocBlockRector implements DowngradeRectorInterface
+{
+}
